requirements.txt
gpt4all
Mastodon.py
python-dotenv